my_fastapi_app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-app"
version = "0.1.28"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
description = "FastAPI CRUD Application"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "pydantic==2.4.2",
    "pydantic-settings==2.0.3",
    "pytest==7.4.3",
    "httpx==0.25.1",
    "alembic==1.12.1"
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
fastapi-app = "app.main:app"